update form field tests
