+++ libs/design/src/atoms/form/form-field/form-field/form-field.component.spec.ts
@@ -3,63 +3,62 @@ import {
   DebugElement,
 } from '@angular/core';
 import {
-  waitForAsync,
   ComponentFixture,
   TestBed,
 } from '@angular/core/testing';
 import {
   ReactiveFormsModule,
   UntypedFormControl,
-  Validators,
 } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 
+import {
+  DAFF_FORM_FIELD_COMPONENTS,
+  DaffFormFieldComponent,
+  DaffFormFieldControl,
+} from '@daffodil/design';
 import { DaffInputComponent } from '@daffodil/design/input';
-
-import { DaffFormFieldComponent } from './form-field.component';
-import { DAFF_FORM_FIELD_COMPONENTS } from '../form-field';
-import { DaffFormFieldControl } from '../form-field-control';
-import { DaffFormFieldMissingControlMessage } from '../form-field-errors';
 
 @Component({ template: `
   <daff-form-field>
     <input daff-input [formControl]="formControl">
-    <daff-error-message></daff-error-message>
+    <daff-hint>Hint></daff-hint>
+    <daff-error-message>Error</daff-error-message>
   </daff-form-field>`,
 imports: [
-  DaffFormFieldComponent,
+  DAFF_FORM_FIELD_COMPONENTS,
   DaffInputComponent,
   ReactiveFormsModule,
 ]})
 
 class WrapperComponent {
-  formControl = new UntypedFormControl('', Validators.required);
+  formControl = new UntypedFormControl();
 }
 
-describe('@daffodil/design | DaffFormFieldComponent | Usage', () => {
+describe('@daffodil/design/form-field | DaffFormFieldComponent | Defaults', () => {
   let wrapper: WrapperComponent;
   let component: DaffFormFieldComponent;
   let fixture: ComponentFixture<WrapperComponent>;
-  let control: DaffFormFieldControl<unknown>;
   let de: DebugElement;
+  let control: DaffFormFieldControl<unknown>;
 
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
       imports: [
         WrapperComponent,
       ],
     })
       .compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(WrapperComponent);
     wrapper = fixture.componentInstance;
-    fixture.detectChanges();
-
     de = fixture.debugElement.query(By.css('daff-form-field'));
     component = de.componentInstance;
     control = component._control;
+
+    fixture.detectChanges();
   });
 
   it('should create', () => {
@@ -72,92 +71,20 @@ describe('@daffodil/design | DaffFormFieldComponent | Usage', () => {
     }));
   });
 
-  describe('error state', () => {
-    it('should set the `daff-error` class on the host element when the child control has an error', () => {
-      wrapper.formControl.markAsTouched();
-      wrapper.formControl.setValue('');
-      fixture.detectChanges();
 
-      expect(wrapper.formControl.errors).toBeTruthy();
-      expect(de.nativeElement.classList.contains('daff-error')).toEqual(true);
-    });
-
-    it('should NOT set the `daff-error` class on the host element when the child control does not have an error', () => {
-      wrapper.formControl.markAsTouched();
-      wrapper.formControl.setValue('Something Valid');
-      fixture.detectChanges();
-
-      expect(wrapper.formControl.errors).toBeFalsy();
-      expect(de.nativeElement.classList.contains('daff-error')).toEqual(false);
-    });
+  it('should have a generated id', () => {
+    expect(component.id).toMatch('daff-form-field-[0-9]*');
   });
 
-  describe('valid state', () => {
-    it('should set the `daff-valid` class on the host element when the child control is valid', () => {
-      wrapper.formControl.markAsTouched();
-      wrapper.formControl.setValue('Something Valid');
-      fixture.detectChanges();
+  it('should have a generated id for the hint', () => {
+    const hint = fixture.debugElement.query(By.css('.daff-form-field__hint-wrapper'));
 
-      expect(wrapper.formControl.valid).toBeTruthy();
-      expect(de.nativeElement.classList.contains('daff-valid')).toEqual(true);
-    });
-
-    it('should NOT set the `daff-valid` class on the host element when the child control is not valid', () => {
-      wrapper.formControl.markAsTouched();
-      wrapper.formControl.setValue('');
-      fixture.detectChanges();
-
-      expect(wrapper.formControl.valid).toBeFalsy();
-      expect(de.nativeElement.classList.contains('daff-valid')).toEqual(false);
-    });
+    expect(hint.nativeElement.id).toMatch('daff-form-field-[0-9]*-hint');
   });
 
-  describe('disabled state', () => {
-    it('should set the `daff-disabled` class on the host element when the child control is disabled', () => {
-      wrapper.formControl.disable();
-      fixture.detectChanges();
+  it('should have a generated id for the error message', () => {
+    const error = fixture.debugElement.query(By.css('.daff-form-field__error-wrapper'));
 
-      expect(wrapper.formControl.disabled).toBeTruthy();
-      expect(de.nativeElement.classList.contains('daff-disabled')).toEqual(true);
-    });
-
-    it('should not set the `daff-disabled` class on the host element when the child control is not disabled', () => {
-      control.state.disabled = false;
-      fixture.detectChanges();
-
-      expect(wrapper.formControl.disabled).toBeFalsy();
-      expect(de.nativeElement.classList.contains('daff-disabled')).toEqual(false);
-    });
-  });
-});
-
-@Component({ template: `
-  <daff-form-field>
-    <daff-error-message></daff-error-message>
-  </daff-form-field>`,
-imports: [
-  DAFF_FORM_FIELD_COMPONENTS,
-]})
-
-class WrapperWithoutControlComponent {}
-
-describe('@daffodil/design | DaffFormFieldComponent | Usage Without Control', () => {
-  let fixture: ComponentFixture<WrapperWithoutControlComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        WrapperWithoutControlComponent,
-      ],
-    })
-      .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(WrapperWithoutControlComponent);
-  });
-
-  it('should throw an error when there is no control present', () => {
-    expect(() => fixture.detectChanges()).toThrowError(DaffFormFieldMissingControlMessage);
+    expect(error.nativeElement.id).toMatch('daff-form-field-[0-9]*-error');
   });
 });
